TINY-8205: Removed extra HTML parse from insert-content's "fix invalid tree structure" logic
